rename Rn.py to rn.py
With Rn instead of rn, doctests fail as shown in issue 3318.
